Changed a pro-error dialog
Instead of ask if we want to quit the game, we must ask if we want to
continue it. This way, the player have just to say "yes" two times to
continue with a saved game, not with a "yes and no" answers where the
player can be wrong and quit the game even if he wouldn't.